Allow retrying remote actions on abrupt failures
This adds logic to treat remote actions that terminate due to a signal as
retryable errors, assuming that such terminations are caused by a worker
crash.

Because this is a hack to paper over a current Remote Build defficiency,
and because this heuristic may be wrong, this feature is hidden behind a
new --snowflake_remote_exit_signals_are_transient_errors flag.

Mitigates #18319.