New C++ linker memory model
This PR instrumets the CppLinkAction to print log lines for every link
operation and dumps the number of inputs, the aggregate size of the
inputs, the predicted memory usage, and the actual consumed memory.

Calculating the aggregate size of the inputs is a costly operation
because we need to expand a nested set and stat all of its files, which
explains why this PR is complex.  I've modified the local memory
estimator to compute the inputs size exactly once per link as we need
this information both to predict memory consumption _and_ to emit the
log line after the spawn completes.

With this, I've run builds of our project with all fastbuild, dbg, and
opt compilation modes, both on amd64 and arm64.  The resulting data
sets showed that input bytes do track memory consumption much better
than input counts, and I've tentatively adjusted the Linux predictior
with a very conservative formula.