Sanitize BEP contents before emitting them
This change adds logic to clean up the contents of the BEP before it
leaves Bazel in an attempt to prevent propagating secrets to a remote
service.

In particular, this implements protections for the following:

*   Secrets stored in the environment which are unconditionally leaked
    via the hidden --client_env flag.

*   Secrets passed via --bes_header, --test_env, --test_arg, and similar
    options.  This is not completely safe to do because a determined
    user could inject secrets into the BEP in other ways.  So... treat
    this as just a safety net for common innocent mistakes.

*   Secrets passed to tools run via "bazel run".  Think, for example,
    doing "bazel run //some/docker:command -- --password 1234".

For the purposes of this change, a "secret" is whatever string we did
not explicitly allow.  This code does not implement secret detection
and it does not indend to, because such code would be based on
heuristics as well.

Note that services like BuildBuddy implement their own redaction of
secrets when they receive the BEP... but it's better if we can strip
them out at the source to minimize exposure.  Defense in depth.