Redact environment in Command debug dumps
Command objects contain the environment variables to pass to them, and their
associated function to print a debug string prints out the environment.

This is not a problem in the common case, but apparently the docker strategy
calls the debug printing functionality on error and leaks the environment
to the console.